Logic + ChartTable
Added the ChartTable method (formerly ChartPanel). Had to rename it becuase it was conflicting with JFreeChart ChartPanel. My chart is up to moving speed, showing all of the required data. I am yet to complete the sorting on the TablePanel. I added the logic and ArrayLists required to create my Aggregate Stats, which will be showed on the JFrame with the chart on it. I also need to figure out how I can implement the details window. I was thinking about putting buttons under each bar on the bar chart that connect to the details window that shows employment specifics. Finally, I need to write out the process in which I added JFreeChart to the module.